Convert getFileCount and getFileSize queries to the Mongo API. Also, map the whole domain package via Morphia, convert a forgotten version field to long (since Morphia requires it) and simplify a few queries.

+++ src/gr/ebs/gss/server/domain/AccountingInfo.java
@@ -46,7 +46,7 @@ public class AccountingInfo  implements Serializable{
 	 */
 	@SuppressWarnings("unused")
 	@Version
-	private int version;
+	private long version;
 
 	/**
 	 * The user whose usage we are noting. We can never change it after
